fundamentals/llm-cli-assistant/requirements.txt
rich
typer
python-dotenv
langchain-openai
langchain-core
